requirements.txt
amqp==5.3.1
annotated-types==0.7.0
ansible==11.6.0
ansible-runner==2.4.1
anyio==4.8.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
black==25.1.0
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
Django==5.0.13
django-allauth==65.6.0
django-anymail==12.0
django-cors-headers==4.7.0
django-cron==0.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.2
langchain==0.3.25
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
lockfile==0.12.2
MarkupSafe==3.0.2
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.78.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pythonping==1.1.4
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resolvelib==1.0.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
svix==1.61.4
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
xxhash==3.5.0
zstandard==0.23.0
